Fixed bad signed type comparison
Use int instead of size_t to suppress signed/unsigned comparison
warning.

+++ source/bxgenbb_help/testing/test_wdecay0.cxx
@@ -17,7 +17,7 @@ int main (int argc_, char ** argv_)
     bool debug = false;
     // bool test  = false;
     bool dump  = false;
-    size_t max_count = 10;
+    int max_count = 10;
 
     int iarg = 1;
     while (iarg < argc_) {
